whisper_demo/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
cmake==3.26.0
cPython==0.0.6
ffmpeg-python==0.2.0
filelock==3.10.0
future==0.18.3
idna==3.3
importlib-metadata==6.1.0
Jinja2==3.1.2
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.21.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
pandas==1.3.4
pycparser==2.21
pymodm==0.4.3
pymongo==3.11.3
python-dateutil==2.8.2
python-telegram-bot==13.4.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.26.0
schedule==1.1.0
six==1.16.0
soundfile==0.11.0
srt==3.5.2
sympy==1.11.1
tiktoken==0.3.1
torch==2.0.0
tornado==6.1
tqdm==4.64.1
triton==2.0.0
typing-extensions==4.5.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vosk==0.3.44
websockets==10.3
zipp==3.15.0
